Wget 1.19 produces WARCs with WARC-Target-URI field values wrapped in angle brackets which does does not follow the WARC spec (see http://lists.gnu.org/archive/html/bug-wget/2017-11/msg00050.html for more details). This causes warcio and projects that depend on warcio (e.g. produce cdx indexes https://github.com/webrecorder/pywb/issues/294) to use the incorrect value for the WARC-Target-URI field.
Added `_ensure_target_uri_format` method to `warcio.recorderloader.ArcWarcRecordLoader` that will detect if the value for the WARC-Target-URI field
contains the Wget bug and corrects it.
Updated `test_archiveiterator.py` to account for this change and added a small WARC (example-wget-bad-target-uir.warc.gz) that is affected by this bug.