Failsafe if it fails to % - encode headers
If Warcio fails to % encode headers, at least when dealing with the exception replace the non-encodeable chracters with a stub. This is not an ideal fix, but at least WARCIO does not fail to decode and encode the same file and the most probable outcome is that a strange cookie gets a few chars replaced.