Handle datetime.utcnow() deprecation
Following the deprecation of utcnow in Python 3.12, we add a
tzinfo timezone argument to date utils. For now this defaults to
None so as not to break comparisons against existing naive
datetimes. If and when naive datetimes are fully deprecated in
a future Python release, we can change the default argument for
tzinfo from None to timezone.utc.